opt: ncoresnm code changes due to typing
Avoid setting or creating attributes of the class which have a
different type (in this case, avoid changing self.par which was a
scalar to a ndarray). This happens in a subclass and we can see that
none of the other subclasses use this approach. There is also the fact
that these classes tend to be created and then immediately used /
deleted so there is no real benefit to storing the values in an
attribute.