opt: drop support for extra args in OptimizerFunc
The OptimizerFunc label is new, but represents parts of the
optimization API. The question is whether there are any extra
arguments other than the first argument. With the statargs/kwargs
changes it is believed that there no-longer is, so support can be
dropped.

Testing shows that the function is always an instance of
sherpa.optmethods.optfcts.Callback, which only supports a
single argument.