TESTS: add explicit numcores tests for optimizers
This should also be tested in the low-level optimizer code paths, but
testing at the UI level does check that the setting is being passed
down (although we can not actually test for that here).

This tests levmar and moncar with numcores=1, 2, and 3. It is a
regression test rather than benig something from first principles, but
the fit results aren't terrible. There are questions over what the
nfev should be, as levmar and moncar report this differently, but
that's separate from this test.

Of note:

- gridsearch is not tested as it is too slow for this dataset
  and I have no experience in setting it up,

- and moncar is interesting as the results depend on OS/architecture,
  including a "swap" of the two lines (which is entirely valid,
  just annoying to deal with).

A simple test of proj() and conf() has also been added with explicit
numcores=1 and 3 cases, but there are definitely more tests of
the numcores-available handling of these cases in the test suite
(as it is the default to use all available cores). Since the tests are
"slow" the numcores=2 has not been added, instead just 1 and 3 are
used (unlike the other tests) added here).