Add the same safeguard to Parameter
As explained in the string in the file itself, we want to be more robust is we combine a parameter with a number. That number might itself be a class that has an `__array_ufunc__` method (if it is, e.g. a np.float64) or it might not (if it is e.g. a Python float). In either case, we want the same output.