Update project.toml for PEP 639
fixes #2247

Address deprecation of "lower-kebab-case" in setup.py in favour of "lower_snake_case".

In version 54 setuptools deprecated naming options in lower-kebab-case, and in version 78 this feature has been removed (thanks to @DougBurke for digging up the relevant part of the changelog: https://setuptools.pypa.io/en/stable/history.html#v78-0-0).

This change is backwards compatible for users of older setuptools version (the naming with an underscore has always been allowed), but might break use scripts that use e.g. sed to monkey with the setting in `setup.cfg` the way we do in the setup of the CI.

Unfortunately there is nothing we can do about it, unless we want to freeze to an older version of setuptools, which has its own drawbacks.

closes #1452