opt: RNG-ification experiment WIP
need to document choices as think going to be diferent for ncoresnm
to ncoresde, since the underlying code choices (e.g. are the
strategies/algorithms hard coded or can they be sent in by the
user).

As part of this:

The MyDifEvo class now creates an explicit RNG rather than allowing it
to be None. This RNG is now explicitly sent to the
ncoresNelderMead.algo entries.

There is some question of whether the RNG should be initialized to a
known state - that is, use a known seed value - or whether it should
be "random" (however the generator decides to be "random"). For now
the initial state is taken from the seed argument, and this is
generally going to be set to the same value.

There are two subtle changes here, where if the rng parameter was
not set then the random seed would be set to the seed, but this
would happen before starting a loop and then within the loop
(thanks to the all_strategies call). This does not seem ideal, and it
was added during #1735 while moving the code to using explicit rather
than implicit RNG.