TESTS: explit RNG for a resample_data test
This test has started to occasionally fail with a statistics
error (while working on RNG changes in the optimizer code base). As
the test is just to check we get no warning messages I have not tried
to find out what is going on, but just decided to use an explicit RNG
to avoid this error.

The error message comes from the statistic calculation, so presumably
the random sampling just happens to hit an edge case somewhere. As
this particular test is not intended to check this feature out I think
it is okay to avoid the problem here (it is a very-rare occurence with
the current code).