XSPEC: support XSPEC 12.15.0
Provide access to the new models in XSPEC 12.15.0 and handle
the change in the rgsxsrc model, which meant that the code would
not build.

This adds support for 17 models:

- additive: XSfeklor, XSvlorentz, XSvvoigt, XSzfeklor, XSzlorentz, XSvlorentz, XSzvoigt, XSzvvoigt

- convolution: XSrgsext

- multiplicative: XSlorabs, XSvlorabs, XSvoigtabs, XSvvoigtabs, XSzlorabs, XSzvorlabs, XSzvoigtabs, XSzvvoigtabs

This build changes the default ATOMDB version from 3.0.9 to 3.1.2.
This can change some fit results, so the tests now explicitly
force the old version. This is not ideal, since at some point version
3.0.9 will disappear, but it's the simplest solution at this time.
The XSPEC interface does provide a better way to set and get this
version, which would allow the change to be made only when needed in
the tests, but this requires changes to the XSPEC module which have
not yet been reviewed.

The user's ~/.xspec/Xspec.init file may need to be removed or updated,
in particular for the AtomDB version, before using the XSPEC models.