scripts: improve update_xspec_functions.py/add_xspec_model.py
The aim of these scripts are to make it easier to update the code when
there is a new XSPEC release. Unfortunately

- update_xspec_functions.py did not handle version-specific macros
well, which means that each time there is a release there needs to be
manual work to add in the necessary '#ifdef' lines, and it ss easy to
get it wrong

- add_xspec_model.py output was designed for XSPEC user models and
not for the XSPEC module, which meant a bunch of edits needed to be
made the output.

* update_xspec_functions.pu

The script has been updated to take the existing settings from the
sherpa/astro/xspec/src/_xspec.cc file, which is currently the only way
we have to get "version" constraints, and then update them for the
new (or changed) models.

This is still not perfect, for example there is no support to properly
handle model names that change or are removed [*], but it should make
it easier to handle upgrades.

[*] so far it looks like XSPEC wants to keep the model names, which
    is why there is no attempt to worry about this possibility

* add_xspec_model.py

The sherpa.astro.utils.xspec code has been updated to support the
"create output for the XSPEC module" case, and the script updated to
use it.