opt: comment out unused ncoresnm/de code
These classes were presumably being developed, but are not actually
used by Sherpa. At least one of them are currently broken, so they
have been commented out. The only place they are used are in
test_opt_original.py and **only** when called directly - i.e. not when
run by pytest - and so they are commented out from the test too.

The code could have been removed, as it can be restored from the git
repository, but for now they have been commented out as it is unclear
what we want to do with the optimisation code.