XSPEC: try to allow module to load with a newer version
There are two ways to define the symbol from `sherpa.astro.xspec._xspec`
that should be called for a model

a) set `_calc` to the symbol
b) set `__function__` to the name of the symbol

The latter has been used to support

- models added in a particular XSPEC version but we still support
  building with an older version of XSPEC,

- and models where the function to call has changed (so older versions
  use one name and newer use another).

However, we can actuallt expand the use of `__function__` so that we use
it for **all** models. This will allow the XSPEC module to load when
the function for a model has been changed but the Sherpa code has not
yet been updated to reflect this (e.g. when XSPEC 12.15.0 came out the
rgsxsrc convolution model was changed to use the C_rgsxsrc C/C++ routine
rather than rgsxsrc FORTRAN routine and so the module failed to load
because `_xspec.rgsxsrc` no-longer existed).  Using this
approach (always setting `__function__` rather than `_calc`) means
that the module can be loaded, even if the `XSrgsxsrc` class can not be
called.