XSPEC: adjust default APEC-related settings
The APEC-like settings added to allow the code to test against
both XSPEC 12.15.0 and earlier versions turned out to need extra
settings. The code  *could* switch the values depending on the
XSPEC version, **but** this would just complicate the tests [*],
and all that we care about is whether it works, not the exact
details (as that's an XSPEC problem).

[*] mainly in the setialization tests.