Shorten implementation of `_make_unop`
In this way, there are no two branches that need to be tested separately. The downside is that the default for `strformat` is repeated from `UnaryOpModel`, which is not good practice either. However, it's not worth the extra complexity in the code to avoid that, give that this is unlikely to be ever changed and any changes would purely be cosmetic on how a model is displayed as a string value, not how it behaves.