opt: better typing of ncoresde.Strategy
The Strategy class

- now has __call__ defined in the base class (as NotImplementedError)
  to show what the subclasses need to support

- the pop argument was set to accept ndarray but it actually is only
  ever called with a SimplexRandom object; this class is set so that
  it has get/set item routines so it behaves like a sequence containing
  ndarray values, which is why the previous type had been ndarray (it
  is possible that using SimplexRandom is a bit too strict, and
  perhaps it should use .opt.SimplexBase, but given this code has not
  changed recently then we go with the actual class)

Given that we know the pop argument is a SimplexBase then we know that
the array access returns a ndarray, so there's no need to use np.array
here. Since we definitely want to copy the trial - as it gets changed
and we do not want this to happen to the SimplexBase data - then we
explicitly call the copy method rather than do a more-implicit copy
with "[:]" (there should be no functional difference here, it is just
more explicit now).