opt: try and standardize calling Opt objects
Add a base `__call__` method to Opt to try and document the calling
convention, but the ncoresde and ncoresnm sub-classes have different
ideas so it is more a guide than a commandment. The ncoresnm version
does have a few changes made to it. Recent changes to mark the calls
as requiring keyword-only arguments have been removed, because of the
attempt to support `*args/**kwargs` arguments, but fortunately the
code has been updated to pass arguments via keyword so any changes
to the argument names (e.g. `tol` to `ftol`) and order are handled
(or caused a test failure and so have been addressed).