opt: rework ncoresde.Strategy to use MyOptOutput
The Strategy class now returns values using the MyOptOutput type
rather than flattening everything into a ndarray. It is not obvious
whether there is a runtime difference here, but the code seems to
be a bit clearer, although

- there is still a need to pack parameters plus the statistic value
  into a ndarray to send to the .opt.SimplexBase routines (changing
  SimplexBase to support this is more-involved change than I wanted
  to support in this commit),

- it might be better to make MyOptOutput a dataclass (or some other
  structured type) rather than accessing elements via position, to
  make the code a little clearer in intent, but again it is not the
  best time to do this.

There is a small amount of changes to related code, such as MyDifEvo,
to go with these changes, and to support very-minor clean up.