opt: very-minor optimizations in SimplexBase
Move centroid calculation after convergence check.

Take advantage of sherpa.utils.random.uniform working on arrays, so we
do not need to loop over each value. This could be extended to a
single call to uniform but it then gets a bit harder to see the logic
of the code so avoid that optimization.

When shrinking, avoid a per-population calculation as it can be done
once. There still needs to be a loop for the statistic
calculation. There is a subtle change in the code here, as it has
changed from looping over [1, self.npar + 1) to [1, self.shape[0]).
It turns out that the only place where shrink is called is set up
to have a "square" simplex, so self.npar + 1 == self.shape[0],
but the change has been made to make the code less "magical".