opt: internal reworking of Simplex code  WIP
Rather than have an array npop by npar + 1, where the last element
of each row is the function value for the parameters of that row,
split out into a simplex array (npop by npar) and fctvals (npop).

This better fits into the other parts of the code, but is it
less efficient? There's also the fact that several tests change,
but why?