opt: internal change to ncoresde.Strategy
The Strategy class, when called, has a similar approach, that is
to loop over the number of parameters and then decide whether to
change a single parameter in the trial distribution based on the
parameter values in a number of other trials/populations.

I have tried to abstract this loop to avoid repitition, but there
is enough differences in the looping that I feel it's not worth it.

However, I think it is a bit clear if we can point out that the
update is basically

    base + factor * delta

and that base and delta are arrays that remain constant, so it's
just the "which parameter is being changed this loop".

In making these changes some of the mathematical operations have
been reworked: they are mathmatically the same - e.g.

    factor * (a - b) + factor * c == factor * (a + c - b)

but they may turn out to be functionally different thanks to
floating-point behavioir (e.g. if a and b are large but of similar
size in the above example). I note that there's no documentation
to say that this may be the case, and we can see that there is
no a-priori reason why a particular ordering is "better" or "worse"
(from the point of floating-point semantics) here given the
random nature of the choices for a/b/c.

This code could technically be a speed improvement, since the
assumption is that, overall, most of the parameters are accessed
so it makes sense to pull out this calculation. However, that has
not been checked and it is secondary to the (assumed) improvement
in documenting the behaviour of the code.