random: improve typing of the uniform routine
Point out that the size argument is optional. The typing documentation
is improving, pointing out some issues like this, as do mypy/pyright.

Address the mode of uniform where arrays of minimum/maximum values are
given. Assume that size=None for this case as it is unlikely to be
set in Sherpa.