In addition to the access by methods, this commit implements access by attribute access. This will be particularly useful for interactive use as the notation is a lot more compact and a lot more readable.
However, one specific class of Models had attribute access for a different purpose before - to define a `FilterModel`. There has long been a remark in the code that `FilterModel` seems unused, but technically it was in the public API, so this commit breaks backwards compatibility. I believe that the attribute access as I want it in the PR and the attribute access for `FilterModel` are mutually exclusive, so backwards compatibility cannot be maintained in this PR.