Set sha256 as hashfunc for cache
Benchmarking indicates that sha265 is a factor of 3-4 faster than md5, and md5 is actually not available on every systems, so there is no reason to use md5 as default. While sha256 might not be cyptographically secure (it has been shown that is can be broken in less time than the key length alone might indicate) collisions are still so unlikely (like 2^63 or something like that) that it's not a concern for this application.