Replace `public` decorator with straight up adding to `__all__`
The datastack modules used a decorator called `public` that would add a decorated function to `__all__` to save the programmer to type it out as a string. That decorator also did some checks to ensure that a function was not added twice, which then cause functions are are aliases to other functions not to be added.
This commit replaces that machinery with direct string access to `__all__`.

fixes #2270