XSPEC: rework how additive models are handled
The XSPEC additive models do not contain a normalization parameter, as
that is handled within the model evaluation by XSPEC. Until now the
Sherpa code would send in a norm parameter and then, in the C++ code,
multiply the model values by the norm value. There is a similar issue
with the table-model code, but there it has not been separated into
spearate classes.

This required "complicating" the macros used to handle the interface
between the C++ module and Python, as well as complicating the Python
code - in particular sherpa.astro.utils.xspec.

So, the norm parameter is now applied within the XSAdditiveModel.calc
call directly. This is somewhat complicated by the recent changes in
PR #2268 which changed XSPEC additive models to apply the norm
parameter in Python scope, sending a value of 1.0 to the compiled
code. Some work will be needed to check what the runtime
characteristics of the new code are.

This will require changes to the handling of XSPEC user models. The
helper routines have been updated so this should be minimal.

As hinted above, XSPEC table models are not handled by this code
(or by the rework in #2268).