XSPEC: rework how additive models are handled
The XSPEC additive models do not contain a normalization parameter, as
that is handled within the model evaluation by XSPEC. Until now the
Sherpa code would send in a norm parameter and then, in the C++ code,
multiply the model values by the norm value. There is a similar issue
with the table-model code, but there it has not been separated into
spearate classes.

This required "complicating" the macros used to handle the interface
between the C++ module and Python, as well as complicating the Python
code - in particular sherpa.astro.utils.xspec.

There is also interest in removing the normalization term from the
Sherpa model cache code, and so reworking how we handle the
normalization might be helpful.

So, the norm parameter is now applied within the XSAdditiveModel.calc
call directly. This has the "interesting" effect of converting the
model evaluation to "always" be np.float64, rather than reflect the
actual model (which, for FORTRAN models, is normally np.float32). The
user still has access to the "original" data. via the model functions
in the sherpa.astro.xspec._xspec module, if they need - but this is
very-low-level.

Code wise it is not clear if it is an improvement, since it just swaps
around where the "is this an additive model" logic needs to be - at
least for the Python code. It does simplify the C++ code a little-bit.

This will require changes to the handling of XSPEC user models. The
helper routines have been updated so this should be minimal.