Based on benchmarking for typical Sherpa use cases, some models are faster to re-calculate than to cache. For those models, the default cache size it now set to 0 to save runtime and memory.
This also adds a "test" that is not run by default (by virtue of having a name that pytest will not recognize) but that can easily be act
ivated by developers as needed.
That test is written for XSPEC model, but similar benchmarking has been done for basic Sherpa models (see # 2258).

fixes #2258