Use `modelCacher1D` (now aliased to `modelCacher`) for 2D models
For a long time, `modelCacher1D` has accepted and arbitrary number of arguments. That is useful because 1D models can be at points (x) or integrated over bins (xlo, xhi). However, that means that is seamlessly also works for (x0, x1) in 2D models or (x0lo, x0hi, x1lo, x1hi) in integrated 2D models. This commit adds an alias `modelCacher = modelCacher1D` (for backwards compatibility we should not remove the old name and I don't see a reason to deprecate it either) and stats using is on some more complex 2D models.