opt: add typing statements to mark callable routines
The main change here is that aliases are created for the functions
used in the sherpa.optmethods modules, and a label for the information
they pass around.

The MyNcores type is updated to explicitly use this in its calc()
call.

Types are added to other parts of this module, primarily the step
argument which can be None or a ndarray. It turns out one of the
SimplexBase classes is always called with step set (and not None), so
an assert is used to let the type checkers know that it is not
set. This could have raised an actual error, but this is low-level
code that users are not expected to be calling, so it was left as an
assert.