Add option `record_steps` to `Fit.fit()` calls
The fit function previously had an option to record statistic and parameter
values for each step of a fit to a file. This is helpful to understand why
a fit may not converge or does not find a global minimum.
However, the file loading and parsing (or capturing in a StringIO object)
is somewhat cumbersome.
This commit adds a simpler way to have that information available in the
`FitResults` object.

By default `record_steps=False` to match previous behavior and conserve memory.

fixes #2269