opt: be explicit when ignoring a value (ncoresde.Strategy)
The Strategy code needs to select n random numbers, which it does
with a call to

    r1, ..., rn = self.init(n)

It turns out that a number of these cases do not use the first
value. We *could* change it to

    r1, ..., rm = self.init(m)  # where m = n - 1

and then update the code to swap the rx values, or maybe just

    r2, ..., rn = self.init(m)

but the problem is that self.init uses the RNG and so the output
will be different if we create n-1 rather than n random numbers.

You would hope the results would still be valid, as there's no
reason why skipping a RNG is meaningful here, but the code has been
tested against the current behaviour and changing it could be
annoying. So the code is now

    _, r2, ..., rn = self.init(n)

which makes it more-obvious that a value is being ignored and does
not create an unused symbol.