ui: internal tweak to get_data/model_contour_prefs
The routines were using "id" when there was no argument called "id",
and so it was using the Python "id" symbol. Which is not correct but
didn't actually result in a problem, because it was used in calls
like "self.get_data_countour(id, recalc=False).contour_prefs" which,
because "recalc=False", ignore the identifier argument.

Originally I had added "id" as an argument to these calls, but at this
time there is no need to do this, since we do not have different
contour classes for different data types (unlike the plot case). If
this ever changes then the code can be re-visited, reviewing the plot
case to see how it should be handled.

This was identified by pyright from the typing statements with a
"why does the type code think I have an object here" moment.