ui: allow multiple dataset identifiers for guess
Add the to_guess method to DataSimulFit to address issue #2284
and take advantage of it in the ui.guess routine.

This is not perfect, since the assumption here is that the multiple
datasets all have the same number of independent axes, and there is no
a-priori reason why this is the case, nor is there an existing check
to enforce this (since technically we can fit different types of data
together). Improvements and additions are left for future work.