ui: allow multiple dataset identifiers for guess  WIP
Add the to_guess method to DataSimulFit to address issue #2284
and take advantage of it in the ui.guess routine.

This is not perfect, since the assumption here is that the multiple
datasets all have the same number of independent axes, and there is no
a-priori reason why this is the case, nor is there an existing check
to enforce this (since technically we can fit different types of data
together). Improvements and additions are left for future work.

Note that the multi-PHA guess case shows it isn't really using the
multiple datasets, probably because the data needs to be re-ordered,
and how do we deal with repeated or overlapping bins?

Yes - some of the guess routines - in sherpa.utils.guess - assume that
the data is ordered (in the indpendent axis; only really checked the
1D cases). So, we could try and re-order all the data, except that the
current interface just gives us ndarray's, so if sent two then there's
no way to know if this is Data1Dint or Data2D. This could - probably,
I haven't thought it through - be fixed by sending around
sherpa.models.regrid.Axis values rather than plain ndarrays, as least
for the current case where we don't have datasets with a mix of
integrated and unintegrated axes (to support something like this we'd
likely want a "container" class to represet each axis).

This is unlikely to be done with our current resources.