Add an interface to ArviZ
ArviZ is a package for the analysis (both plotting and statistical computations) of Baysian models. It has a number of functionalities and default plots that are useful for MCMC chains that we don't have in Sherpa, and thus it is useful to provide an interface to pass Sherpa MCMC results into ArviZ.
This PR defines the interface functions, undates the docs and optional dependencies to include ArviZ. One of the CI runs is modified to install it as a dependency.

It also includes a slight reformating of the print statements in the rst docs to make
them pass as doctest.

fixes #1942