xspec: improve error handling
Use a more-modern C++ idiom for checking a string is not empty.

In doing this I realised there were no Python checks of what happens
if abundance + table values are invalid. This is because the code is
written such that most users can not actually cause this to happen. So
the test code has to call the C++ code directly (that is, it calls a
sherpa.astro.xspec._xspec routine that a user is not expected to
call).

This then led me to realize that I had not cleared the stream handling
if the getAbundance call threw a NoInitializer error, so this has been
added.

Address typos found during review.

Co-authored-by: wmclaugh <wmclaugh@cfa.harvard.edu>