Update http -> https
All servers are hand-checked at least once and in a number of cases the servers or specific pages do no longer exist. In those cases, a best available replacement link is used.

No changes are applied to "extern" or where the copyright of files places limits on how they can be modified.
Similarly, comments in C++ code are generally not updated and for comments in Python code, judgment is used e.g. versioneer, which is just vendored from an external project) is not touched.

closes #2172

[skip ci]