Improve docs for hard and soft limits
Hard and soft limits are already explained in the docs, but some details that might be useful for developers extending or building on Sherpa have been added. Since the audience of the added information are developers, and this is not typically an area that users need ot care about, the information has been added to a module docstring. I also made some of the relevant module level constants public - a developer might want to reset them.

closes #1480