build: ask Python for the platform library name
In testing a build out on docker, using Fedora 42, I found that
it was using the "lib64" and not "lib" directory. At the time the
choice is made the directories do not exist, so an existance
check is not enough. It also feels like using the Python sys
setting would make the most sense, but I am not convince (given
how things work in Python), so let us see. Of course, it is not
for all library directories.