Allow optimizer options to be set on initialization
We already had two ways to set options for an optimizer:

opt = Optimizer()
opt['par'] = value
opt.par = value

however, both require one extra line per option and in optimigic, we will need to set something (at least the name of the algorithm) every time it's used, so that is a bit tedious. This commit adds a way to pass in arguments on the creation of the optimizer.

opt = Optimizer(par=value)

closes #150
closes #580
closes #611
closes #783
closes #1476