Allow restoring of BinaryOpModels
`CompositeModel` (one of the parent classes of `BinaryOpModel`) gains an attribute `parts` in their `__init__` function. However, when they are restored from a pickeled save, `__init__` is never called. Eventually, the `parts` attribute is taken from the pickle and added to the objects dict, but we can't access it before that is done.

fixes #2301

I believe that this regressions was introduced in #2166, but I did not actually bisect it because that didn't seem necessary after I understood what causes the problem.

This commit also adds tests for the regression and a minor syntax clean-up that helped me with interactive debugging.