Make optimization tests with montecarlo reproducable
I found that there is one and only one test (test_optmethods/test_Func1) that is flaky with the montecarlo. It fails about 1/3 of the time both on this feature branch and on main on my machine, though given the involvement of random numbers that might differ between numpy implementations.
(We haven ot see much of that, probably because these tests are relatively new in production).

In principle, it is good to run all the tests with varying seeds so that, over time, a larger parameter space is tried. However, for this particular test, the failure rate is too high to be practical. So, I'm fixing the seed for all montecarlo tests.