TESTS: add --runcores flag to run extra multi-core tests
From review by @hamogu, only run multiple "multi-core" tests when
asked to, to reduce the run-time requirements of the tests (since
they can be quite slow). This lets several tests which were labelled
with a comment of "we don't test with numcores=2 to save time" to
allow this test to be made.