Restore Callback class but mark as deprecated
Just in case anyone has used the Callback class, add in an alias
so that they will end up using StatCallback. However, mark it as
deprecated, and make sure a user calling it will see a message.
Since DeprecationWarning messages are not easily seen, we use a
log message instead to make sure the user sees it (unless they
have not set up the logger, which is possible). As it is unlikely
anyone is using this symbol I did not want to spend much time in
coming up with a solution.

The version-added label is updated to 4.18.0 since this PR did
not make it for 4.17.1.