opt: add commentary to ncoresde.Strategy class
The Strategy call method is documented, mainly to say "I have
experimented with moving some of the code around because it is
cleaner and may be faster, but it can subtly change the results,
so I have not applied these changes here" in case any follow-up
work is done.

The code changes are to make the operations consistent within the
sub-classes, and to hide the use of self.rng so it is easier to see
where the random-ness is being used. There should be no changes to
the results because of these changes.