estmethods: move local lambda functions to top level
Move some of the local routines for the confidence call to the
module level. This is to support use of methods other than fork
for multiprocessing.

Some typing statements are added, a few f-strings were introduced, and
a minimal amount of documentation, as well as trying to indicate the
relationship of these routines to the confidence call (by changing the
names to add a "c_" prefix).