estmethods: rework the worker function for confidence
Previously a local lambda was used to create the function that is
called to estimate the per-parameter confidence errors. This routine
is called in both the serial and parallel cases.

To allow the routine to be defined at the top-level, to support
non-fork multiprocessing methods, it has been moved into a class.
This allows sending in all the information the method needs while
retaining the simple interface for actually calling it.