fit: move error estimation call-backs into top-level clesses
Move a number of lambda routines to top-level classes so that
they can be run when multiprocessing is not set to fork.

The classes could have been made dataclasses but the typing arguments
are not well defined yet, so leave them as normal classes. They can
be changed to a dataclass in the future.