opt: internal change to grid_search
Move the statistic callback from a local lambda to a class in the
module, to support multiprocessing with methids other than fork.

Comments have been added to identify possible improvements with the
logging (either drop it or move it to use the actual Sherpa logging
instance), as it is not the time to work on that.