opt: minor internal clean up
Rework the "fdjac" routines so that they pass around "structued" data,
rather than stick everything into a single array and then have to do
type conversions. As part of this, send around the actual value rather
than the index to it that is then used to reconstruct the value.

Take advantage of array broadcasting to avoid two lines. There are
possible future improvements to this routine - such as NumPy clipping
- but leave that for a later time. This is also in the fdjac code.

In lmdif the type checker thought that covar could be used before
being set. This should in fact be impossible, but the over-writing of
the "internal" info value to convert it to the "external" value makes
it hard to see. This is fixed by creating an explicit "external" info
value (info2). However, this still does not appease the type
checker/linter (as the logic is hard to follow) so explicitly set the
covar value to None and then just check it has been set before its
use (as the code would have blown up before with a NameError I feel
confident of just using an assert here).