CI: add a test-run with multiprocessing=spawn (conda)
Extend one of the conda runs to re-run the tests with a
different multiprocessing method ("spawn") rather than "fork".

Unfortunately this really slows down the tests, so it is applied to
just the macOS ARM run, as this is one of the fastest builds (whilst
still testing a large number of optional components).

As with the ncores=1 for pip change, the idea is to re-run the tests
for a build. The code is somewhat messy, but it is not worth cleaning
up until there is a need.