CI: try to exercise ncores=1/multiprocessing=spawn in tests
The idea is to add these tests to existing runs, rather than
create a new build for them. The selection of where to apply
these is somewhat random and dependent on how long the tests
run. The approach taken is a bit hap-hazard and could be cleaned
up, but that is only worth it if we want more of these tests.