utils: explicit imports for the ui module
The use of sherpa.all has, over the years, annoyed me, as it makes
it hard to track down what is being used, and linting tools do not
seem to like it. So, move to explicitly import the modules rather
than rely on sherpa.all, which is somewhat un-pythonic.

However, we do not remove sherpa.all since it does provide the
user with information about what optional parts of the system - such
as the plotting backend - are available. There may be a "better"
approach, but it does not seem worth changing it here.