TESTS: add examples of bug #2314
The sherpa.ui.xxx_sample routines should be repeatable given
the same RNG. We currently have two ways of doing this:

a) globally, with np.random.seed with the default ui.set_rng
   value of None

b) locally, by calling ui.set_rng with a RNG.

Check that we get the same behaviour with both approaches.

The aim is to remove option a at some distant point.