utils: ensure record_steps can be sent to the fit. Fix #2318
Rework the check for valid arguments to make it less hard-coded
(so, going back to how the code used to be written) but also
supporting the new keywords that are supported.

As part of this, update the documentation.  For CIAO we need the text
to be before the example, as there's no other way to break up each
example (without a lot of tedious trial and error). This also fixes up
some spacing.

Ensure the versionchanged lines are in descending version order, to
match the rest of the system (this is also needed for the ahelp
generation).