utils: tweak examples for fit and the record_steps option
For CIAO we need the text to be before the example, as there's no
other way to break up each example (without a lot of tedious trial
and error). This also fixes up some spacing.

Ensure the versionchanged lines are in descending version order,
to match the rest of the system (this is also needed for the ahelp
generation).