mark sherpa.astro.utils._region as supporting free threading
Since it is not clear whether the region library is thread safe,
mark all use with a global mutex. This can hopefully be relaxed
or removed. The same mutex is used to mark the DM call, but it
is not clear if this is sufficient, as it is a different library
and can be called from other code (this is only an issue if
building with the CIAO Data Model code).

Converting this from single- to multiple-phase initialization is
not attempted here.