CI: add a test-run with ncores=1 (pip)
In looking at the multiprocessing code it is obvious that there are
potentially-subtle differences in the code paths when the ncpus is 1
rather than None (as nowadays most machines, even on CI, have multiple
CPUs).

So, make sure one test is run with the "ncores: 1" setting in the
configuratuion file (sherpa.rc) rather than the default of "ncores:
None".

There are a number of ways this could be done, such as adding an extra
run of the tests after a successful standard run, which has the
advantage of ensuring "the best" test coverage (since coverage can not
easily handle the multiprocessing code) but at the expense of
time/computation.

This is only done for the pip case at present.

As part of this, simplify test_lrt/test_lrt_multicore in test_sim.py,
as this can be done with a single parametrized test rather than two
essentially-identical tests. However, the previous test included
checks for multi/ncpus which appear to no-longer be needed. Is this
because the code has changed - to somehow allow the "ncores > 2"
version to use two generators even when no multiprocessing is
selected - or because there is a subtle bug in the multi-core case
when only a single core is requested?

Note that without this test change the tests would fail with the
new ncores=1 CI run.