CI: add a test-run with ncores=1 (pip)
In looking at the multiprocessing code it is obvious that there are
potentially-subtle differences in the code paths when the ncpus is 1
rather than None (as nowadays most machines, even on CI, have multiple
CPUs).

Rather than add a new test, just for ncores=1, add an extra step to
the pip run to re-run the tests with ncores=1. Only one run is changed
to do this extra run (one of the runs with most of the extra packages
installed). The current approach is rather messy but it does not feel
worth it to try to abstract this at this time.

Possibilities include:

- dropping the "ncores: None" test when ncores=1, but I think it is
  useful to check the same code passes when the ncores setting is
  changed

- add something similar to the conda run, but I am leaving that as
  I plan to add a conda run where the multiprocessing method can
  be changed, and I do not want to add too many extra CI runs.

As part of this, simplify test_lrt/test_lrt_multicore in test_sim.py,
as this can be done with a single parametrized test rather than two
essentially-identical tests. However, the previous test included
checks for multi/ncpus which appear to no-longer be needed. Is this
because the code has changed - to somehow allow the "ncores > 2"
version to use two generators even when no multiprocessing is
selected - or because there is a subtle bug in the multi-core case
when only a single core is requested?

Note that without this test change the tests would fail with the
new ncores=1 CI run.